+++ NEWS.md
@@ -1,5 +1,9 @@
 # User visible changes for EasyFITS
 
+## Version 0.5.3
+
+- Fix writing in a file with a given header and image array. Header is now merged with success.
+
 ## Version 0.5.2
 
 - Call `write(file::FitsFile,FitsImageHDU{T},dims...)` to create a new FITS image HDU